Move the konnectivity pod check to the common inttest package
So it can be reused by other integration tests.
